+++ src/main/java/blackjack/domain/card/HoldCards.java
@@ -1,7 +1,5 @@
 package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
