+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,12 +1,9 @@
-package blackjack;
+package blackjack.domain.card;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Number;
-import blackjack.domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class CardTest {
