+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,9 +1,5 @@
-package blackjack;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Deck;
-import blackjack.domain.card.Suit;
-import blackjack.domain.card.Number;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
