+++ src/test/java/blackjack/domain/card/NumberTest.java
@@ -1,14 +1,14 @@
-package blackjack;
+package blackjack.domain.card;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import blackjack.domain.card.Number;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
 import java.util.List;
 import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class NumberTest {
     @ParameterizedTest
